fix(time): 修复时间解析中的整数溢出问题
- 将整数乘法转换为长整型乘法以防止溢出
- 在 parseTime 方法中添加类型转换确保计算准确性
- 避免因整数溢出导致的时间计算错误

+++ src/main/java/com/honzel/core/util/time/TimeRangeUtils.java
@@ -605,7 +605,7 @@ public class TimeRangeUtils {
     }
 
     private static LocalTime parseTime(int seq) {
-        return LocalTime.MIN.plusMinutes(seq * TIME_UNIT_IN_MINUTES);
+        return LocalTime.MIN.plusMinutes((long) seq * TIME_UNIT_IN_MINUTES);
     }
 
     /**
